requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bert-score==0.3.12
betty-ml==0.2.0
black==21.12b0
bleach==5.0.0
blinker==1.6.2
blis==0.7.7
Brotli==1.0.9
cachetools==5.2.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.2.0
colorama==0.4.4
constantly==15.1.0
contourpy==1.0.7
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.6
datasets==2.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
emoji==2.2.0
entrypoints==0.4
evaluate==0.1.2
exceptiongroup==1.1.0
executing==0.8.3
fastapi==0.89.1
fastjsonschema==2.15.3
filelock==3.7.0
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.5.0
gcloud==0.17.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.46.3
gym==0.21.0
h11==0.14.0
hkdf==0.0.3
httplib2==0.21.0
huggingface-hub==0.8.1
humanize==4.4.0
hyperlink==21.0.0
icecream==2.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.12.0
incremental==21.3.0
inflate64==0.3.1
iniconfig==2.0.0
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-contrib-core==0.4.0
-e git+https://github.com/ipython-contrib/jupyter_contrib_nbextensions.git@2248f6e15753d0eb4339215313ae21ba09cbdf74#egg=jupyter_contrib_nbextensions
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
jupyterthemes==0.20.0
jws==0.1.3
kiwisolver==1.4.2
langcodes==3.3.0
lesscpy==0.15.0
lightning-utilities==0.6.0.post0
llvmlite==0.38.1
lxml==4.9.1
magic-wormhole==0.12.0
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
multivolumefile==0.2.3
murmurhash==1.0.7
mypy-extensions==0.4.3
nbclient==0.6.0
nbconvert==7.0.0
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numba==0.55.2
numpy==1.23.3
oauth2client==3.0.0
oauthlib==3.2.0
openai==0.27.6
opencv-python==4.5.5.64
packaging==21.3
pandas==1.5.0
pandas-stubs==1.2.0.58
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
plotly==5.13.0
pluggy==1.0.0
ply==3.11
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
promptsource==0.2.3
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.4.3
pycryptodomex==3.17
pydantic==1.8.2
pydeck==0.8.0
pyDeprecate==0.3.2
pyflakes==3.0.1
pyglet==1.5.21
Pygments==2.15.1
Pympler==1.0.1
PyNaCl==1.5.0
pynvim==0.4.3
PyOpenGL==3.1.5
pyOpenSSL==22.1.0
pyparsing==3.0.8
pyppmd==1.0.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.2.1
python-cfonts==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jwt==2.0.1
pytorch-lightning==1.9.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.3
regex==2022.3.15
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
responses==0.18.0
rich==13.3.5
rouge-score==0.1.2
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.9.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spake2==0.8
srsly==2.4.3
stack-data==0.2.0
starlette==0.22.0
streamlit==1.22.0
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
texttable==1.6.7
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.1
torchvision==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
transformers==4.19.2
Twisted==22.8.0
txaio==22.2.1
txtorcon==22.0.0
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.20.0
validators==0.20.0
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.4
Wikipedia-API==0.5.4
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
